requirements.txt
google-generativeai>=0.3.0
flask>=2.0.1
python-dotenv>=0.19.0
requests>=2.26.0
pytest>=6.2.5
google-adk